requirements.txt
streamlit
openai
langchain
serpapi
google-search-results
# numexpr
boto3